# sources/antimage-dev/internal/app/api/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "antimage API",
    "version": "0.3.0",
    "description": "Go-native antimage panel API for dashboard, bot, and third-party client developers.\n\nAuthentication: the dashboard uses an opaque server-side session stored in the HttpOnly `antimage_session` cookie. Session permissions are loaded from the database on every request, so permission changes take effect without logging out. Bots and external integrations should use personal `rk_...` API keys as `Authorization: Bearer <api_key>`. Legacy JWT creation at `POST /admin/token` and `POST /api/admin/token` remains available only for older Marzban-compatible clients.\n\nDo not automate username/password login for bots. API keys can be revoked independently and inherit the owner admin permissions, service scope, and limits.\n\nMutation behavior: user, service, node, config, and host mutations are transactional where applicable. Runtime changes are synchronized to nodes through gRPC and recorded for recovery."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current antimage panel origin"
    }
  ],
  "security": [
    {
      "cookieAuth": []
    },
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/settings/subscriptions": {
      "get": {
        "tags": [
          "Settings"
        ],
        "summary": "Get subscription settings",
        "description": "Return global subscription settings, templates metadata, aliases, ports, and admin overrides.\n\nAuthentication: send `Authorization: Bearer <api_key>` for bots and third-party integrations. Create personal admin API keys from My Account -> API Keys; each key inherits the owner admin permissions and limits. The current dashboard JWT from `POST /api/admin/token` is also accepted for interactive dashboard sessions and for Swagger UI when opened from inside the panel.\n\nPermissions: Panel read requires an authenticated admin. Settings mutation, backup, templates, and TODO/410 settings routes require sudo/full_access.",
        "responses": {
          "200": {
            "$ref": "#/components/responses/JSON"
          }
        },
        "operationId": "getSettingsSubscriptions"
      },
      "put": {
        "tags": [
          "Settings"
        ],
        "summary": "Update subscription settings",
        "description": "Update global subscription settings, aliases, ports, template directories, and client format flags.\n\nAuthentication: send `Authorization: Bearer <api_key>` for bots and third-party integrations. Create personal admin API keys from My Account -> API Keys; each key inherits the owner admin permissions and limits. The current dashboard JWT from `POST /api/admin/token` is also accepted for interactive dashboard sessions and for Swagger UI when opened from inside the panel.\n\nPermissions: Panel read requires an authenticated admin. Settings mutation, backup, templates, and TODO/410 settings routes require sudo/full_access.\n\nRequest body: send JSON unless this operation explicitly declares multipart/form-data. Unknown fields are generally ignored only when the underlying handler is compatibility-oriented; business-rule violations return 400/403/409.\n\nSide effects: this is a mutation. Database writes are transactional where applicable; runtime-affecting changes enqueue node operations or call the node controller as documented by the route.",
        "requestBody": {
          "$ref": "#/components/requestBodies/JSON"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/JSON"
          }
        },
        "operationId": "putSettingsSubscriptions"
      }
    },
...
    "/sub/{identifier}": {
      "get": {
        "tags": [
          "Subscriptions"
        ],
        "summary": "Render subscription",
        "description": "Render a user subscription by token, key, username/key pair, or legacy identifier.\n\nAuthentication: public route; no bearer token is required.\n\nPermissions: Public endpoint. Do not send admin bearer tokens unless you are using a custom integration that deliberately does so.\n\nRoute details: Returns config text, JSON, YAML, HTML, or client-specific output depending on identifier, client suffix, User-Agent, and subscription settings.\n\nOutput notes: subscription responses set client/profile headers such as subscription-userinfo when available and may return plain text, base64, JSON, YAML, or HTML depending on format resolution.",
        "security": [],
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Subscription response in a client-selected format."
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        },
        "operationId": "getSubIdentifier"
      }
    },
...
    "/sub/{part1}/{part2}": {
      "get": {
        "tags": [
          "Subscriptions"
        ],
        "summary": "Render two-part subscription path",
        "description": "Render either identifier/client_type or username/credential_key subscription paths. antimage resolves the meaning from stored user credentials and known client types.\n\nAuthentication: public route; no bearer token is required.\n\nPermissions: Public endpoint. Do not send admin bearer tokens unless you are using a custom integration that deliberately does so.\n\nRoute details: Ambiguous two-part subscription paths are resolved by checking whether the second part is a known client type or a credential key. Known client types include `v2ray`, `v2ray-json`, `clash`, `clash-meta`, `sing-box`, `outline`, `openvpn`, `wireguard`/`wg`, `happ`, `v2raytun`, `throne`, `shadowrocket`, `karing`, `hiddify`, `clash-mi`, `incy`, `passwall`, and `nekobox`.\n\nOutput notes: subscription responses set client/profile headers such as subscription-userinfo when available and may return plain text, base64, JSON, YAML, or HTML depending on format resolution.",
        "security": [],
        "parameters": [
          {
            "name": "part1",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "part2",
            "in": "path",
            "required": true,
            "description": "Credential key, legacy subscription key, or known client type such as `v2raytun`, `karing`, `hiddify`, `shadowrocket`, `wireguard`, `wg`, or `openvpn`.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Subscription response."
          }
        },
        "operationId": "getSubPart1Part2"
      }
    },
...
    "/sub/{username}/{credential_key}/{client_type}": {
      "get": {
        "tags": [
          "Subscriptions"
        ],
        "summary": "Username/key client subscription",
        "description": "Render client-specific subscription output using explicit username and credential key.\n\nAuthentication: public route; no bearer token is required.\n\nPermissions: Public endpoint. Do not send admin bearer tokens unless you are using a custom integration that deliberately does so.\n\nRoute details: Use this form when your integration knows username, credential key, and desired client format explicitly. Known client types include `v2ray`, `v2ray-json`, `clash`, `clash-meta`, `sing-box`, `outline`, `openvpn`, `wireguard`/`wg`, `happ`, `v2raytun`, `throne`, `shadowrocket`, `karing`, `hiddify`, `clash-mi`, `incy`, `passwall`, and `nekobox`.\n\nOutput notes: subscription responses set client/profile headers such as subscription-userinfo when available and may return plain text, base64, JSON, YAML, or HTML depending on format resolution.",
        "security": [],
        "parameters": [
          {
            "$ref": "#/components/parameters/Username"
          },
          {
            "name": "credential_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "client_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "v2ray",
                "v2ray-json",
                "clash",
                "clash-meta",
                "sing-box",
                "outline",
                "openvpn",
                "wireguard",
                "wg",
                "happ",
                "v2raytun",
                "throne",
                "shadowrocket",
                "karing",
                "hiddify",
                "clash-mi",
                "incy",
                "passwall",
                "nekobox"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Subscription response."
          }
        },
        "operationId": "getSubUsernameCredentialKeyClientType"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "antimage admin API key or legacy JWT",
        "description": "Use a personal rk_ API key for bots and third-party integrations. Legacy JWTs from /admin/token remain accepted for older Marzban-compatible clients."
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "antimage_session",
        "description": "Opaque HttpOnly dashboard session cookie issued by POST /api/auth/login. It is managed by the browser and is not readable by dashboard JavaScript."
      }
    },
    "parameters": {
      "Username": {
        "name": "username",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "antimage username. Matching is case-insensitive in most user lookup paths."
      },
...
    },
    "requestBodies": {
      "JSON": {
        "required": false,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Object"
            },
            "example": {
              "name": "example",
              "status": "active",
              "settings": {}
            }
          }
        },
        "description": "Generic JSON request body. See each operation description for required business fields and side effects."
      },
...
    },
    "responses": {
      "JSON": {
        "description": "JSON response. Exact object fields follow the dashboard-compatible antimage API response for the route.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/AnyValue"
            },
            "example": {
              "id": 1,
              "status": "ok",
              "detail": "Operation completed"
            }
          }
        }
      },
...
      "Error": {
        "description": "Error response. Most handlers return `{ \"detail\": \"...\" }`; some compatibility routes may return a plain text body when handled by net/http helpers.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "schemas": {
...
      "Object": {
        "type": "object",
        "additionalProperties": true,
        "example": {
          "name": "example",
          "status": "active",
          "settings": {}
        }
      },
      "AnyValue": {
        "nullable": true,
        "example": {
          "id": 1,
          "status": "ok"
        },
        "oneOf": [
          {
            "type": "object",
            "additionalProperties": true
          },
          {
            "type": "array",
            "items": {}
          },
          {
            "type": "string"
          },
          {
            "type": "number"
          },
          {
            "type": "integer"
          },
          {
            "type": "boolean"
          }
        ]
      },
      "Error": {
        "type": "object",
        "properties": {
          "detail": {
            "type": "string"
          }
        }
      },
...
    }
  }
}